+++ app/src/main/java/org/ssg_tab/domain/repository/quiz/QuizRepository.kt
@@ -1,6 +1,6 @@
 package org.ssg_tab.domain.repository.quiz
 
-import org.ssg_tab.domain.model.entity.QuizEntity
+import org.ssg_tab.domain.model.entity.quiz.QuizEntity
 
 interface QuizRepository {
     suspend fun getQuizList(categoryId: Int, difficulty: String): Result<List<QuizEntity>>
