+++ app/src/main/java/org/ssg_tab/presentation/ui/guide/state/QuizContract.kt
@@ -1,7 +1,7 @@
 package org.ssg_tab.presentation.ui.guide.state
 
 import androidx.compose.runtime.Immutable
-import org.ssg_tab.domain.model.entity.QuizEntity
+import org.ssg_tab.domain.model.entity.quiz.QuizEntity
 
 @Immutable
 data class QuizContract(
@@ -9,7 +9,7 @@ data class QuizContract(
     val currentQuizIndex: Int = 0,
     val selectedAnswer: Int = -1,
     val isLoading: Boolean = false,
-    val error: String? = null
+    val error: String? = null,
 ) {
     val currentQuiz: QuizEntity?
         get() = quizList.getOrNull(currentQuizIndex)
